+++ src/NewGroupFolder.h
@@ -26,7 +26,7 @@
 	IBOutlet NSTextField * folderName;
 	IBOutlet NSButton * saveButton;
 	IBOutlet NSButton * cancelButton;
-	int parentId;
+	NSInteger parentId;
 }
 
 @property(strong) NSArray * topObjects;
@@ -36,5 +36,5 @@
 -(IBAction)doCancel:(id)sender;
 
 // General functions
--(void)newGroupFolder:(NSWindow *)window underParent:(int)itemId;
+-(void)newGroupFolder:(NSWindow *)window underParent:(NSInteger)itemId;
 @end
